// prisma/generated/zod/inputTypeSchemas/CategorySumOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const CategorySumOrderByAggregateInputSchema: z.ZodType<Prisma.CategorySumOrderByAggregateInput> = z.object({
  id: z.lazy(() => SortOrderSchema).optional(),
  maxPartySize: z.lazy(() => SortOrderSchema).optional(),
  maxParties: z.lazy(() => SortOrderSchema).optional(),
  competitionId: z.lazy(() => SortOrderSchema).optional()
}).strict();

export default CategorySumOrderByAggregateInputSchema;